requirements.txt
aiofiles==22.1.0
appdirs==1.4.4
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Color-Console==10.0
colorama==0.4.5
config==0.5.1
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
docopt==0.6.2
fake-useragent==0.1.11
filelock==3.8.0
Flask==2.2.2
h11==0.13.0
heroku==0.1.4
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
parse==1.19.0
pipenv==2022.9.4
pipreqs==0.4.11
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
pyee==8.2.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==1.5
python-dotenv==0.21.0
python-telegram-bot==12.8
python2==1.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
sanic==22.6.2
sanic-routing==22.3.0
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.21.0
telegram==0.0.1
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.12.1
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
uvloop==0.16.0
virtualenv==20.16.4
virtualenv-clone==0.5.7
w3lib==2.0.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
zipp==3.8.1
